Move get_scheduler and read slots from settings
